+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Fix tests.
+- External route not working on storefront.
 
 ## [2.7.1] - 2019-03-01